new update to alerts to have nws
